% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_dada.R
\name{simple_dada}
\alias{simple_dada}
\title{simple_dada}
\usage{
simple_dada(path, paired = TRUE, cores = 0)
}
\arguments{
\item{path}{Directory for project. Raw reads should be 
contained in a sub-folder named raw_reads}

\item{paired}{Whether or not the program should look
for forward (R1) and reverse (R2) fastq files.}

\item{cores}{The number of CPU cores/threads to use.}
}
\value{
data.table
}
\description{
Runs through the DADA2 routine with dynamic paramter
assignment. In situations where the data is clean this
will work well. If the data is not clean, it will be 
bet to either run the DADA2 routine. or each
sub-function.
}
